pyproject.toml
[project]
name = "multi-ai-agents-using-adk-travel-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.23.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
